+++ src/translator/JsTranslator.js
