+++ src/translator/Translator.js
@@ -1,0 +1,8 @@
+import { NodeVisitor } from 'ast';
+
+export default class Translator extends NodeVisitor {
+    async translate(astRoot) {
+        await this.init();
+        await this.visit(astRoot);
+    }
+};
